Restructure project with src/ directory
