+++ pages/_app.js
@@ -1,32 +1,13 @@
-
-
-import * as React from "react";
-import * as MainLayout from "../layouts/MainLayout.js";
-import * as Router from "next/router";
-
 import 'styles/main.css'
-;
+import { make as ResApp } from "src/App.js"
 
-function $$default(props) {
-  var router = Router.useRouter();
-  var content = React.createElement(props.Component, props.pageProps);
-  var match = router.route;
-  if (match === "/examples") {
-    return React.createElement(MainLayout.make, {
-                children: null
-              }, React.createElement("h1", {
-                    className: "font-bold"
-                  }, "Examples Section"), React.createElement("div", undefined, content));
-  } else {
-    return React.createElement(MainLayout.make, {
-                children: content
-              });
-  }
+// Note:
+// We need to wrap the make call with
+// a Fast-Refresh conform function name,
+// (in this case, uppercased first letter)
+//
+// If you don't do this, your Fast-Refresh will
+// not work!
+export default function App(props) {
+  return <ResApp {...props}/>;
 }
-
-export {
-  $$default ,
-  $$default as default,
-  
-}
-/*  Not a pure module */
