--- pages/_app.resi
@@ -1,2 +1,0 @@
-type props
-let default: props => React.element
