--- pages/index.resi
@@ -1,1 +1,0 @@
-let default: unit => React.element
