+++ src/App.resi
@@ -1,0 +1,2 @@
+type props
+let make: props => React.element
