+++ src/Index.resi
@@ -1,0 +1,1 @@
+let make: unit => React.element
